commit 20/04

+++ src/controllers/jogos.controllers.js
@@ -1,38 +1,54 @@
 const JogosService = require('../services/jogos.services');
+const mongoose = require('mongoose');
 
-const findJogosController = (req, res) => {
-  res.send(JogosService.findJogoService());
+const findJogosController = async (req, res) => {
+  const Alljogos = await JogosService.findJogoService();
+  res.send(Alljogos);
 };
 
-const findJogoByIdController = (req, res) => {
-  if (!req.params.id) {
-    return res.status(400).send({ message: 'Id não encontrado!' });
+
+
+
+const findJogoByIdController = async (req, res) => {
+
+  if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+    res.status(400).send({ message: 'ID inválido!' });
+    return;
+  }
+
+  jogoEscolhido = await JogosService.findJogoByIdService(req.params.id);
+
+  if (!jogoEscolhido) {
+    return res.status(400).send({ message: 'Jogo não encontrado' });
   } else {
-    if (!JogosService.findJogoByIdService(req.params.id)) {
-      return res.status(400).send({ message: 'Jogo não encontrado' });
-    } else {
-      res.send(JogosService.findJogoByIdService(req.params.id));
-    }
+    res.send(jogoEscolhido);
   }
 };
 
-const addJogoController = (req, res) => {
+
+
+
+const addJogoController = async (req, res) => {
 
   let retorno = JogosService.addJogoService(req.body);
 
   if (
-    !req.body.id ||
+    !req.body ||
     !req.body.nome ||
-    !req.body.imagem ||
-    !req.body.preço ||
-    !req.body.estoque
+    !req.body.genero ||
+    !req.body.plataforma ||
+    !req.body.estoque ||
+    !req.body.lançamento ||
+    !req.body.preco
   ) {
     return res.status(400).send({
       message:
         'Você não preencheu todos os dados para adicionar um novo jogo a seleção!',
     });
   } else {
-    if (retorno == 'ok') {
+    if (retorno) {
+      const newJogo = await JogosService.addJogoService(req.body);
+
       res.send({ message: 'Jogo cadastrado com sucesso!' });
     } else {
       res.status(400).send({ message: 'Houve um erro' });
@@ -40,33 +56,52 @@ const addJogoController = (req, res) => {
   }
 };
 
-const updateJogosController = (req, res) => {
 
-  
 
-  if (!req.body || !req.body.nome || !req.body.preco || !req.body.imagem || !req.body.estoque) {
-
-    return res.status(400).send({ message: "Você não preencheu todos os dados para editar o jogo!" });
+const updateJogosController = async (req, res) => {
+  if (
+    !req.body ||
+    !req.body.nome ||
+    !req.body.genero ||
+    !req.body.plataforma ||
+    !req.body.estoque ||
+    !req.body.lançamento ||
+    !req.body.preco
+  ) {
+    return res.status(400).send({
+      message: 'Você não preencheu todos os dados para editar o jogo!',
+    });
   }
 
-  if (!JogosService.findJogoByIdService(req.params.id)) {
-
-    return res.status(400).send({ message: 'Jogo não encontrado' });
-
+  if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+    res.status(400).send({ message: 'ID inválido!' });
+    return;
   }
 
-  const updatedJojo = JogosService.uptadeJogoService(+req.params.id, req.body);
-  res.send(updatedJojo);
+  const updatedJogo = await JogosService.uptadeJogoService(
+    req.params.id,
+    req.body,
+  );
+  res.send(updatedJogo);
 };
 
-const deleteJogoController = (req, res) => {
 
-    if (!JogosService.findJogoByIdService(req.params.id)) {
 
-        return res.status(400).send({ message: 'Jogo não encontrado' });
-     }
 
-  JogosService.deleteJogoService(req.params.id);
+const deleteJogoController = async (req, res) => {
+  if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+    res.status(400).send({ message: 'ID inválido!' });
+    return;
+  }
+
+  const jogoEscolhido = await JogosService.findJogoByIdService(req.params.id)
+
+
+  if (!jogoEscolhido) {
+    return res.status(400).send({ message: 'Jogo não encontrado' });
+  }
+
+  await JogosService.deleteJogoService(req.params.id);
 
   res.send('Personagem deletado com sucesso');
 };
